🚀 백엔드 기능 배포

+++ movieinfo/views.py
@@ -2,56 +2,60 @@ import json
 import re
 import datetime
 
-# 테스트를 위한 csrf exempt
-
-from django.shortcuts import render
 from django.db.models import Q
 from rest_framework import viewsets
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework.response import Response
 
 from . import utils
-from .models import (
-    MovieInfo,
-    OneLineCritic,
-    GPTAnalysis,
-    TestLikeMovie,
-    TestWatchedMovie,
-    TestWatchlistMovie,
-)
+from .models import MovieInfo, OneLineCritic, GPTAnalysis
+from .permissions import OneLineCriticIsAuthorOrReadOnly
+from accounts.models import User
+from accounts.models import LikeMovie, WatchedMovie, WatchlistMovie
+
 from .serializers import (
     MovieInfoSerializers,
     OneLineCriticSerializers,
+    OneLineCriticSaveSerializers,
     OneLineCriticCreateUpdateSerializers,
     GPTAnalysisSerializers,
-    TestLikeMovieSerializers,
-    TestWahtchlistMovieSerializers,
-    TestWatchedMovieSerializers,
+    LikeMovieSerializers,
+    WatchlistMovieSerializers,
+    WatchedMovieSerializers,
 )
 from .detail_summary_with_GPT import send_data_to_GPT as GPT
 
 
-class SerachMovieAPIView(APIView):
+class SearchMovieAPIView(APIView):
     def post(self, request):
         query = json.loads(request.body)["query"]
         query = re.sub(" ", "", query)
-        res = utils.getMovieInfo(query)
+        res = utils.get_movie_info(query)
 
         if res.status_code == 200:
             data = res.data
             if data["Data"][0]["Count"] == 0:
                 return Response({"message": "검색 결과가 없습니다."})
-            utils.saveMovieInfo(data)
+            utils.save_movie_info(data)
 
         queryset = MovieInfo.objects.filter(Q(searchTitle__icontains=query))
         serializer = MovieInfoSerializers(queryset, many=True)
         return Response(serializer.data)
 
 
+class MovieInfoViewSet(viewsets.ModelViewSet):
+    queryset = MovieInfo.objects.all()
+    serializer_class = MovieInfoSerializers
+    http_method_names = ["get"]  # allowed methods
+
+
 class OneLineCriticViewSet(viewsets.ModelViewSet):
     queryset = OneLineCritic.objects.all()
     serializer_class = OneLineCriticSerializers
+    permission_classes = [OneLineCriticIsAuthorOrReadOnly]
+    http_method_names = ["get", "post", "delete", "patch"]
 
     def list(self, request, *args, **kwargs):
         pk = self.kwargs["movie_id"]
@@ -60,7 +64,10 @@ class OneLineCriticViewSet(viewsets.ModelViewSet):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def create(self, request, *args, **kwargs):
-        serializer = OneLineCriticCreateUpdateSerializers(data=request.data)
+        data = request.data
+        data["author"] = User.objects.get(username=request.user).pk
+        data["movie"] = self.kwargs["movie_id"]
+        serializer = OneLineCriticSaveSerializers(data=data)
         serializer.is_valid(raise_exception=True)
         self.perform_create(serializer)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -77,10 +84,13 @@ class OneLineCriticViewSet(viewsets.ModelViewSet):
         req = self.request.data
         content = req["content"]
         starpoint = req["starpoint"]
+        user = User.objects.get(id=req["author"])
+        movie = MovieInfo.objects.get(id=self.kwargs["movie_id"])
         serializer.save(
             content=content,
             starpoint=starpoint,
-            movie=MovieInfo.objects.get(id=self.kwargs["movie_id"]),
+            movie=movie,
+            author=user,
         )
 
     def retrieve(self, request, *args, **kwargs):
@@ -108,10 +118,12 @@ class OneLineCriticViewSet(viewsets.ModelViewSet):
         pk = self.kwargs["pk"]
         movie_id = self.kwargs["movie_id"]
         try:
-            instance = OneLineCritic.objects.get(Q(id=pk) & Q(movie__id=movie_id))
+            instance = OneLineCritic.objects.get(Q(id=pk) & Q(movie=movie_id))
             data = request.data
+            data["user"] = request.user
+            data["movie_id"] = movie_id
             serializer = OneLineCriticCreateUpdateSerializers(
-                instance=instance, data=data, partial=True
+                instance=instance, data=data
             )
             serializer.is_valid(raise_exception=True)
             self.perform_update(serializer)
@@ -127,16 +139,22 @@ class OneLineCriticViewSet(viewsets.ModelViewSet):
         serializer.save(
             content=content,
             starpoint=starpoint,
-            movie=MovieInfo.objects.get(id=self.kwargs["movie_id"]),
         )
 
     def destroy(self, request, *args, **kwargs):
-        return super().destroy(request, *args, **kwargs)
+        pk = self.kwargs["pk"]
+        movie_id = self.kwargs["movie_id"]
 
-
-class MovieInfoViewSet(viewsets.ModelViewSet):
-    queryset = MovieInfo.objects.all()
-    serializer_class = MovieInfoSerializers
+        try:
+            instance = OneLineCritic.objects.get(Q(id=pk) & Q(movie=movie_id))
+            if request.user != instance.author:
+                raise ValueError
+        except:
+            errorMessage = {"message": "잘못된 응답입니다."}
+            return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
+        instance.delete()
+        message = {"message": "삭제되었습니다."}
+        return Response(message, status=status.HTTP_204_NO_CONTENT)
 
 
 class UserLWWViewSet(viewsets.ModelViewSet):
@@ -145,25 +163,51 @@ class UserLWWViewSet(viewsets.ModelViewSet):
 
     """
 
-    queryset = TestLikeMovie.objects.all()
-    serializer_class = TestLikeMovieSerializers
-    # authentication_classes = ()
+    queryset = LikeMovie.objects.all()
+    serializer_class = LikeMovieSerializers
+    permission_classes = [IsAuthenticated]
+    lookup_field = "user"
 
     def list(self, request, *args, **kwargs):
         pk = self.kwargs["movie_id"]
         mode = kwargs["mode"]
 
         if mode == "like":
-            queryset = TestLikeMovie.objects.filter(movie__id=pk)
-            serializer = TestLikeMovieSerializers(queryset, many=True)
+            queryset = LikeMovie.objects.filter(movie__id=pk)
+            serializer = LikeMovieSerializers(queryset, many=True)
             return Response(serializer.data, status=status.HTTP_200_OK)
         elif mode == "watchlist":
-            queryset = TestWatchlistMovie.objects.filter(movie__id=pk)
-            serializer = TestWahtchlistMovieSerializers(queryset, many=True)
+            queryset = WatchlistMovie.objects.filter(movie__id=pk)
+            serializer = WatchlistMovieSerializers(queryset, many=True)
             return Response(serializer.data, status=status.HTTP_200_OK)
         elif mode == "watched":
-            queryset = TestWatchedMovie.objects.filter(movie__id=pk)
-            serializer = TestWatchedMovieSerializers(queryset, many=True)
+            queryset = WatchedMovie.objects.filter(movie__id=pk)
+            serializer = WatchedMovieSerializers(queryset, many=True)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        else:
+            errorMessage = {"message": "잘못된 응답입니다."}
+            return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
+
+    def retrieve(self, request, *args, **kwargs):
+        movie_id = self.kwargs["movie_id"]
+        user = self.kwargs["user"]
+        mode = kwargs["mode"]
+
+        if mode == "like":
+            queryset = LikeMovie.objects.get(Q(movie__id=movie_id) & Q(user__pk=user))
+            serializer = LikeMovieSerializers(queryset)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        elif mode == "watchlist":
+            queryset = WatchlistMovie.objects.get(
+                Q(movie__id=movie_id) & Q(user__pk=user)
+            )
+            serializer = WatchlistMovieSerializers(queryset)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        elif mode == "watched":
+            queryset = WatchedMovie.objects.get(
+                Q(movie__id=movie_id) & Q(user__pk=user)
+            )
+            serializer = WatchedMovieSerializers(queryset)
             return Response(serializer.data, status=status.HTTP_200_OK)
         else:
             errorMessage = {"message": "잘못된 응답입니다."}
@@ -172,20 +216,28 @@ class UserLWWViewSet(viewsets.ModelViewSet):
     def create(self, request, *args, **kwargs):
         movie_id = kwargs["movie_id"]
         mode = kwargs["mode"]
+        user = request.user
+        data = {"movie_id": movie_id, "user": user}
         if mode == "like":
-            instance = TestLikeMovieSerializers(data={"movie_id": movie_id})
+            instance = LikeMovieSerializers(data=data)
             instance.is_valid(raise_exception=True)
-            instance.save(movie=MovieInfo.objects.get(id=movie_id))
+            instance.save(
+                movie=MovieInfo.objects.get(id=movie_id), user=self.request.user
+            )
             return Response(instance.data, status=status.HTTP_201_CREATED)
         elif mode == "watchlist":
-            instance = TestWahtchlistMovieSerializers(data={"movie_id": movie_id})
+            instance = WatchlistMovieSerializers(data=data)
             instance.is_valid(raise_exception=True)
-            instance.save(movie=MovieInfo.objects.get(id=movie_id))
+            instance.save(
+                movie=MovieInfo.objects.get(id=movie_id), user=self.request.user
+            )
             return Response(instance.data, status=status.HTTP_201_CREATED)
         elif mode == "watched":
-            instance = TestWatchedMovieSerializers(data={"movie_id": movie_id})
+            instance = WatchedMovieSerializers(data=data)
             instance.is_valid(raise_exception=True)
-            instance.save(movie=MovieInfo.objects.get(id=movie_id))
+            instance.save(
+                movie=MovieInfo.objects.get(id=movie_id), user=self.request.user
+            )
             return Response(instance.data, status=status.HTTP_201_CREATED)
         else:
             errorMessage = {"message": "잘못된 응답입니다."}
@@ -194,38 +246,49 @@ class UserLWWViewSet(viewsets.ModelViewSet):
     def destroy(self, request, *args, **kwargs):
         movie_id = kwargs["movie_id"]
         mode = kwargs["mode"]
-        if mode == "like":
-            try:
-                instance = TestLikeMovie.objects.get(
-                    movie=MovieInfo.objects.get(id=movie_id)
-                )
-                instance.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
-            except:
-                errorMessage = {"message": "잘못된 응답입니다."}
-                return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
-        elif mode == "watchlist":
-            try:
-                instance = TestWatchlistMovie.objects.get(
-                    movie=MovieInfo.objects.get(id=movie_id)
-                )
-                instance.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
-            except:
-                errorMessage = {"message": "잘못된 응답입니다."}
-                return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
-        elif mode == "watched":
-            try:
-                instance = TestWatchedMovie.objects.get(
-                    movie=MovieInfo.objects.get(id=movie_id)
-                )
-                instance.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
-            except:
-                errorMessage = {"message": "잘못된 응답입니다."}
-                return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
+        user = User.objects.get(pk=kwargs["user"])
+        if request.user == user:
+            if mode == "like":
+                try:
+                    instance = LikeMovie.objects.get(
+                        Q(user=user) & Q(movie=MovieInfo.objects.get(id=movie_id))
+                    )
+                    if user == instance.user:
+                        instance.delete()
+                        return Response(status=status.HTTP_204_NO_CONTENT)
+                    else:
+                        raise ValueError
+                except:
+                    errorMessage = {"message": "잘못된 응답입니다."}
+                    return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
+            elif mode == "watchlist":
+                try:
+                    instance = WatchlistMovie.objects.get(
+                        Q(user=user) & Q(movie=MovieInfo.objects.get(id=movie_id))
+                    )
+                    if user == instance.user:
+                        instance.delete()
+                        return Response(status=status.HTTP_204_NO_CONTENT)
+                    else:
+                        raise ValueError
+                except:
+                    errorMessage = {"message": "잘못된 응답입니다."}
+                    return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
+            elif mode == "watched":
+                try:
+                    instance = WatchedMovie.objects.get(
+                        Q(user=user) & Q(movie=MovieInfo.objects.get(id=movie_id))
+                    )
+                    if user == instance.user:
+                        instance.delete()
+                        return Response(status=status.HTTP_204_NO_CONTENT)
+                    else:
+                        raise ValueError
+                except:
+                    errorMessage = {"message": "잘못된 응답입니다."}
+                    return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
         else:
-            errorMessage = {"message": "잘못된 응답입니다."}
+            errorMessage = {"message": "권한이 없습니다."}
             return Response(errorMessage, status=status.HTTP_400_BAD_REQUEST)
